regalloc: Implement graph coloring allocator
Summary:
Implement a Chaitin-Briggs style allocator with special handling for range instructions and for storing params at the end of the frame.

It still needs tuning and improvement, as evidenced by the number of XXX's and TODO's in the code, but the essential structure is there.

While at it, I fixed the handling of check-cast; {D5035274} was wrong. We should insert move instructions before a check-cast, not after. The register allocator will be responsible for ensuring that the move instruction doesn't clobber live values. See the comment above `find_check_cast` for details.

Differential Revision: D5113824

fbshipit-source-id: 2966beb8d62f34d2d3122107a6c1f48b3d4cd6d9