Introduce Switch Inline pass
Summary:
This pass allows us to mitigate some of the overhead
created by the fact that our codegenerators operate
in 2 steps (local, global).
As a result sometimes the code that will be generated is
not known at the time of the local compilation and is
only generated at the global step.

In this instance the DI framework only knows what the
factory method of a specific binding will be at the
global step (due to Asserted bindings and Multibinds)
so it never explicitly
generates a specific method call when a binding is
requested, but instead uses a polymorphic static call
to a get method that can return all DI enabled objects.
(This is used also ti improve the local lib compilation
efficiency)

At the same time time the framework needs to be
able to refer to any binding by a simple int id to be
able to generate an efficient version of the Lazy and
Provider proxy objects.

At redex time however there are many cases where those
static calls to UL.factorymap.get(id, ...) can be optimized
and inlined because the id is static and known before hand.

This pass utilizes the infra from SimpleInline and combines it
with a understanding of the structure of these switchmaps
to eliminate the overhead of this indirection.

Differential Revision: D5064122

fbshipit-source-id: 9fa2aa1635304847bf59242ec28b4a4a05f293b1