Cleanup Switch
Summary:
This pass allows us to mitigate some of the overhead
created by the monolithic ultralight switch and is
similar to the old CleanupProvidersPass that
performed the same task for the old DI structure.

In essence after SwitchInline and dead code elimination
there are many cases where we know that certain ids
of the switch map are never used because calls to
the switchmap can acquire ids in a certain number of
ways. (sget to the id class or UL.id.id(const))

This pass utilizes the understanding of the structure of
these switchmaps to eliminate unused refs

Differential Revision: D5064265

fbshipit-source-id: 1b50ddce0971327369b5166867df7fbef08adca4