Make sure we are calling `invoke_static_to_direct`
Summary:
As title.

The inliner's public api provides 2 ways to inline:
- inline_methods()
- inline_callees()

In the first case, it calls `invoke_static_to_direct()`, but for the second
case, it doesn't. At the same time, it doesn't make sense to call it after each
`inline_callees()` call, given we are going to call it multiple times (as in
the case of the builder pass)

This way we make sure that the user doesn't have to call
invoke_direct_to_static at the end. (Otherwise, we can also go with that).

Reviewed By: int3

Differential Revision: D5170646

fbshipit-source-id: b1d278366b4ec5dcda35dea9484660f5cad89cf4