Allow uninitialized fields with the reg allocator
Summary:
This fixes a TODO from D5136101.

Given now we allow over 16 registers, we should consider cases where the reg is >= 16.

When creating the null instruction we consider both the register and the type of the field:
- if is wide: use OPCODE_CONST_WIDE
- if the reg is < 16: OPCODE_CONST_4
- if the reg is >= 16: OPCODE_CONST_16

Also, we are allocating a new register for each of the uninitialized fields.
Given that, we might see a small change in stats bot, since now the reg allocator is not enabled and some of the builders might not be removed any more because of the reg number getting > 16.

If the difference is big (though I don't expect it to be) we can either land this with the reg allocator enabling or reconsider allocating a new register for each unitialized field.

Reviewed By: int3

Differential Revision: D5139724

fbshipit-source-id: 1f4bbde5236ed9495985cb3163b5b78d7a90189d