track the time of each pass
Summary:
add a metric for each pass that records the amount of time that pass took

It might be noisy, so take it with a grain of salt. We also want to use
this to figure out how noisy the time is in sandcastle.

Later, we should add a metric for the entire running time of redex.

Reviewed By: emma0303

Differential Revision: D5202078

fbshipit-source-id: 1520f195063654eaa2e02ef84a3314a9acdbc66d