Intra-Method RedundantMoveElimination (diff 1 of 2)
Summary:
Right now, RME knows nothing when it enters a block
after this diff (and the next) it will know more,
so it can delete more stuff.
And we all like deleting stuff :D

AliasedRegisters is implemented by a graph. Nodes are register values and
edges represent an alias between them.

I implement join as the union of the edges and meet as the intersection
of edges. It's very similar to a SetAbstractDomain.

Differential Revision: D5168327

fbshipit-source-id: bf3fef2e5aa6495c80baaba990e07f1f509da1d6