Adds a new Redex pass for reordering Interface list for each Class.
Summary:
There is a linear walk over the list of interfaces to resolve a call to an Interface which involves some overhead. This diff is the first attempt to reorder the list of Interfaces so that the linear walk terminates quicker.

The run is timed on my laptop with light load. Breakdown is here: {F67546221}

Reviewed By: dariorussi, minjang

Differential Revision: D5180692

fbshipit-source-id: 30d56a0862b5dd87d2be5b02eb8fd8c1a964e237