Add a timer section to statsbot
Summary:
add a "time_stats" section to redex-stats.txt

It has a list of all the timers we created in the order that they finished.

See also D5206485

revert the previous stacked diff that only measured the time for the passes.

Also, these times should be taken with a grain of salt. Expect some noise from the sandcastle machines (we're one of 3 processes running at once on the same machine)

Reviewed By: int3

Differential Revision: D5205647

fbshipit-source-id: 68d7d91ec070d20378fddb2473f4927255facca6