Add MethodMerger
Summary:
MethodMerger is intend to merge fragment model getters to a single (at least for now) switched statement and update all getter call sites to invoke the merged switch dispatch instead.
It is an infra piece for the upcoming TypeErasure change.
In this diff, MethodMerger simply tries to merge all fragment model reference field getters.
I plan to merge the code change but not enable it.

Reviewed By: dariorussi

Differential Revision: D5188941

fbshipit-source-id: c8378dea4c5cf9a7d7a929e90cb852712a64999d