Adapt constant propagation to be able to work with intra-procedural const-prop analysis (2/3)
Summary:
This should be almost a no-op modulo a slight regression:  This code does almost the same as the previous constant propagation package but it's even dumber: it doesn't propagate constants through basic-blocks and it doesn't get clever trying to reach a fix point and propagating constants over and over, instead it only does it once and only at the basic block level.

This diff is not intended to land alone but it's useful since conceptually this is the next logical step to get intra-procedural constant propagation working.  The next diff will be to integrate this with the fix point constant-prop analysis checked in early and see if we get any wins.

Reviewed By: justinjhendrick

Differential Revision: D5187337

fbshipit-source-id: 9078595bc9f070efc7f99790001ab0d5aed0b8f4