Keep lower number of null regs
Summary:
As title.

Previously, we were creating more registers than needed per method. With this
change, we will allocate max of 4 null registers, depending on the type: if it
is wide / or not, and if it is primitive / or not.

No other change, except the reduced number of null regs allocated.

Reviewed By: int3

Differential Revision: D5229437

fbshipit-source-id: e9d13d8001a9179cddd783e54cd385f4d3145352