More on type system and class/virtual scopes
Summary:
This is an extension and definition of the API around the type system.
If someone that undrestands C++ wants to take a look at the API and make suggestions to fix my poor C++ that would be good.
Of course, please, keep useless details to yourself but I think I can use some help to drive this better.
Unit tests will improve in time but I have dome some decent ones so happy for now.

build[automation_fbandroid_redexunstable,automation_messenger_redexunstable]

Reviewed By: thezhangwei

Differential Revision: D5203822

fbshipit-source-id: 3af3c4ea984d7aaf2f558d3f12711b9ad41b51f5