Definition of type erasure set via config and some fixes
Summary:
Changed the config driving type erasure to contain the set of base type for the type to erase. This is similar to what we had for GQL but extended to CS and it should all be a bit less brittle.
Also fixed ClassScope to contain the interface scope for a type that defines interfaces.
And finally fixed method distribution across mergers.

Reviewed By: thezhangwei

Differential Revision: D5273319

fbshipit-source-id: 46e6fdb5c6ae64005699258815ed0b6853e6e676