Make CleanupFeature invalidate ul ids when deleting them
Summary:
We want to maintain the following invariant so that subsequent passes see a consistent world.
If we see an SGet to the ultralight id class then that id has a valid binding within the map.

Cleanup Feature invalidates a bunch of cases within the factorymap by converting them to null,
but it doesn't cleanup the sgets.
This diff converts the sgets to ids that CleanupFeatures invalidates to const/4 -1.
This basically corresponds to an invalid id which throws a RuntimeException from the factorymap

Reviewed By: kgsharma, dariorussi

Differential Revision: D5263682

fbshipit-source-id: 79a177d9d6884b473cdcf6e100daca5c594b6089