Introduce a pass plugin mechanism into redex
Summary:
We would like to customize the behavior of some passes based on the presence
of other passes. This allows the system to implement additional capabilities
depending on the specific configurations.

This diff introduces the PluginRegistry were every pass can introduce a plugin
interface and a pass specific registry.
Other passes can then register polymorphic construction functors which should be
called during run_pass to construct plugin implementations for the pass run to
delegate information/calls to.

Reviewed By: minjang

Differential Revision: D5263772

fbshipit-source-id: 5c1f391ec37969670ef4cdb4d9588458f702d02b