Enable plugin support into InterDexPass
Summary:
The InterDex pass performs a core operation that is inherently extentable.
Introduce plugin hookups during the important operations so that the
InterDex pass can be extended in interesting ways.

For example:
A plugin can decide to generate and include new DexClasses into the secondary
Dex based on the dex classes present in the dex.
A plugin can decide to not emit a class to the dex (usually because the pass
registering the plugin will ensure that all refs to the class will be
eliminated)

Reviewed By: int3

Differential Revision: D5263992

fbshipit-source-id: bea2f198317c0c27ef24aec207d029d93cc60922