Rewrite WorkQueue to use std::thread
Summary: This commit re-implements WorkQueue to use std::thread and extends it to handle more general workloads.  The plan is to use this structure for multithreading within a single optimization pass, such as Peephole or RemoveGotos, which does not have any dependencies between classes, but might need some initialization work done on a per-thread basis.

Reviewed By: int3, minjang

Differential Revision: D5268060

fbshipit-source-id: 9deefbc539737b3292a506957b84ed616e08ab44