Don't treat builders that are passed to the buildee's constructor as escaping
Summary:
If the buildee's constructor accepts the builder as an unique argument, then
don't treat that case as escaping (at the same time we check that the new
method that we will create does not exist)

For the cases where the buildee accepts the builder as the only argument, we
create a new constructor, that will take all the builder's fields as arguments.

Precondition:
* in the constructor, the builder doesn't change registers (for simplicity)

Changes:
  invoke-direct {<v_buildee>, <v_builder>}
  ->
  iget <v_field_1>, <v_builder>
  iget <v_field_2>, <v_builder>
  ...
  invoke-drect {<v_buildee>, <v_field_1>, <v_field_2> ...}

Added this change only for the case where the register allocator is enabled.
Based on this, updated the tests to reflect the state with the register
allocator on.

Differential Revision: D5234605

fbshipit-source-id: b7b1e6d740d4c394c2e5e0dd4263c94c15a31a91