Add coalescable interference edges between wide operands
Summary:
We seem to have uncovered a bug in the ART interpreter's handling of unaligned wide registers. In particular, given the following code:

  xor-long v1, v0, v9

v1 does *not* equal to `v0 ^ v9` after this operation. The overlapping src and dest registers are likely the source of the problem. This diff ensures that we never create these partial overlaps of the src and dest registers of a wide instruction.

Note that this is fairly similar to the move-wide problem: while we don't want to create partial register overlaps within a single instruction, we still want to be able to coalesce those registers. As such I've reused (and renamed) the mechanism that suree404 implemented to mark these edges specially and ignore them when doing coalescing.

Differential Revision: D5300760

fbshipit-source-id: f74c28731a2971e8d062245f542f491eb011bec4