remove-builders: Don't reuse the null reg for different primitive types
Summary:
This leads to a verification error. In particular, this seems to be a problem:

  const-wide v0 #0
  invoke-static {v0, v1} Foo.bar(J)V
  invoke-static {v0, v1} Foo.baz(D)V

Because after the first invoke, the verifier treats v0 as a double type.

As far as I can tell, null regs for objects aren't specialized to a particular
class type. I.e.

  const-wide v0 #0
  invoke-static {v0} Foo.bar(Ljava/lang/String;)V
  invoke-static {v0} Foo.baz(Ljava/lang/Integer;)V

Seems to be fine. So we just need to worry about distinguishing the primitive
types, and this diff mostly does that. It only distinguishes between int-like
types and floats, as well as between longs and doubles. I think we may still
have problems if we encounter integers of different widths reusing the same
null reg...  but that's a fix for another day, I'm tired.

I also renamed 'null reg' to 'zero reg' since that's what the Dalvik verifier
calls the types of these zero-initialized registers.

Reviewed By: emma0303

Differential Revision: D5309325

fbshipit-source-id: acb0b19a3cf1cde79db7facf7fc86ea249be2129