dce: No need to prune CFG edges before removing unreachable blocks
Summary: The initial implementation of DCE tried to prune the CFG edges after removing dead code. However, it did so imperfectly and resulted in the CFGs getting out of sync with the underlying code. {D5091222} just rebuilds the entire CFG to be on the safe, more-likely-to-be-correct side. However, I'd forgotten to delete the old pruning code when I committed that change, so this diff finishes the job.

Reviewed By: justinjhendrick

Differential Revision: D5178994

fbshipit-source-id: 2cfbeb2f44573a92bb6e0c6f666eed34bea0ab43