create DedupBlocksPass
Summary:
This pass removes blocks that are duplicates in a method.

If a method has multiple blocks with the same code and the same successors,
delete all but one of the blocks. Naming one of them the canonical block.

Then, reroute all the predecessors of all the blocks to that canonical
block.

Reviewed By: int3

Differential Revision: D5119803

fbshipit-source-id: ea0eee02ee2d17b988c40c675772f92260d2fba3