Implement PatriciaTreeMapAbstractEnvironment
Summary:
Started on this in order to speed up the live range analysis in regalloc. "Unfortunately" it seems that mmey's parallelization work has made this much less impactful than it would otherwise have been, but I figured I'd finish it up anyway.

Mostly just cribbed the code from PatricaTreeSet.h, made it into a map, and made the resulting code pass the tests for the HashedAbstractEnvironment.

Differential Revision: D5324606

fbshipit-source-id: fbfececf750edb3aa879ae6ab0e3c365899a61b7