Fixed a bug in the fixpoint iterator that only manifests itself when analyzing self-loops
Summary:
This is a rather obscure bug that only manifests itself when analyzing a basic block with a self-loop (in Dex code, that only happens for infinite or do-while loops). This bug occurs when the default-constructible value of the abstract domain is not the identity for the join_with operation when applied to non-_|_ values. Hence, that bug is invisible when performing liveness analysis or any set-based analysis (the default-constructed value of an abstract set is the empty set), but may affect the precision (not the correctness) of analyses that use abstract environments (where the default-constructed value is Top).

I uncovered this problem while working on the pointer analysis. It is highly unlikely that this bug has a significant impact on existing analyses, since it only affects rare loop constructs.

Reviewed By: minjang

Differential Revision: D5362587

fbshipit-source-id: dee5ae7d196252e8eb6386bb60dc7ba1b5c82d70