Fix instrumentation testing bugs
Summary: This commit fixes the failing instrumentation tests.  The issue was that redex was optimizing away the constructor because it wasn't used, since JUnit uses reflection to call the constructor.  Therefore, JUnit complained that no constructor was available, even for objects with the default constructor.  By adding a new test case that creates each object, we circumvent this issue.

Reviewed By: justinjhendrick

Differential Revision: D5366205

fbshipit-source-id: 464193d990d77079994a7ba6eb43369fb81229d0