Fix and improve and test StripDebugInfo pass
Summary:
StripDebugInfo stopped working after the FatMethod changes went in because the debug info moved from being int the DexCode object, to be in the FatMethod object in the IRCode objects. This method was disabled at some point after it stopped working. This change fixes the StripDebugInfo pass, but leaves it disabled in the fb4a.config file. The plan is to enable it when we need a file size win next quarter.

I added both integration tests to ensure that each individual setting works as expected and also added instrumented tests to verify end to end redex runs work as expected.

Reviewed By: int3, justinjhendrick

Differential Revision: D5377764

fbshipit-source-id: 410803ae4d511cb43a4cd9fe10524d08175a5ebe