Capture workqueue lambda by value
Summary: An issue opened on Github describes a stacktrace involving a null function pointer that was being invoked in WorkQueue.  This seems like it is an issue with capturing the stack variable by reference, since the stack variable will go out of scope.  std::function's are copyable, we can just copy by value instead.

Reviewed By: maxmg22, minjang, int3

Differential Revision: D5393799

fbshipit-source-id: db1d7986a1d2a7fc2ad00a0f838fc264304c1b1d