IRInstruction::size should not speculatively convert to 2addr
Summary: Currently, when querying the size of an IRInstruction, we check if it can be converted to 2addr form (i.e. it must have dest == src(0), and be an instruction like add-int, mul-int, etc). If so, we return the size of the instruction in 2addr form, even if the value of the m_opcode field is something like `add-int` and not `add-int/2addr`. This is weird, unintuitive behavior that can lead to bugs.

Reviewed By: justinjhendrick

Differential Revision: D5380671

fbshipit-source-id: cbbddda7af61774130982809b09e9b0924fb6462