Make it possible to call edge() on a non-const CFG
Summary:
Right now, it's not possible to call `CFG::edge()` if the CFG is non-const.
This is because overload resolution happens before access checking, so such a
call will resolve to the non-const `edge` method first, before realizing that
it is private and therefore an access error.

This diff renames the private overload of `edge` so that it is no longer an
issue.

Reviewed By: justinjhendrick

Differential Revision: D5485980

fbshipit-source-id: bb4ba4d27337085fcda248abc7204c29e126d490