Usability improvements and bug fix in dex.py
Summary:
I was using `dex.py` and had to go look in the source for the
format of a class_filter. Also, I forgot to give it files and was confused
by the lack of output (oops). Then I tried to give it an apk and it crashed.

So, I made things a little clearer for the next person.

Also, I fixed a bug in the printing of protos. A proto_idx was being used
to index into the type table, but they have a separate table
https://source.android.com/devices/tech/dalvik/dex-format#proto-id-item

Reviewed By: clayborg

Differential Revision: D5493709

fbshipit-source-id: f77fe88d97a14a693af915b095c933588105a9ba