Disable buildee constructor update (same as D5502999, but for unstable)
Summary:
Porting this to unstable too.

There was a bug with games loading (no error, nor
verification error thrown, but games weren't loading). Disabling the pass fixed
it for the moment.

Enabling and only disabling the buildee constructor replacement also fixes the
issue. Added this feature under a config param and disabled for now, until I
send a proper fix for it.

Reviewed By: dariorussi

Differential Revision: D5503034

fbshipit-source-id: 7eabc962078159d5b34009a69dac5ea4d2092956