Add boolean value test for HashedAbstractEnvironment
Summary: The assertion that the abstract environment is a value is found within methods of this class, but it is difficult to check outside of the class because of the templates.  This method can be used to first make sure the map is initialized before the values are enumerated over.

Reviewed By: minjang

Differential Revision: D5525764

fbshipit-source-id: 3a6d17ab8cc2ad0f4daab3634f525614bf8a0984