Fix nondeterminism in sync()
Summary:
IRCode::sync() was producing nondeterministic output because `encode_offset`
could create MethodItemEntries without initializing their `addr` field. The
loop that called `encode_offset` would then loop over these
partially-initialized MIEs with unpredictable results.

This diff fixes it while doing a bunch of refactoring so that things are more
obviously correct. Instead of storing the address as a field in the MIE, we
store it in a separate map. That way, we can skip over the MIEs without
addresses assigned. Attempting to access them also becomes a runtime exception
instead of silently creating nondeterminism. Also, this is a little cleaner:
only `balloon` and `sync` have use for addresses -- the individual passes don't --
so the field was useless for them.

Since balloon() already had a addr-to-entry map, it seemed to make sense to
make it a bidirectional map instead of having a separate entry-to-addr map.
However, this requires us to have a 1-to-1 mapping. Due to our removal of
no-ops and pseudo-opcodes in balloon(), this was not possible. We were mapping
more than one address to a given entry for nops, and we were mapping all
pseudo-opcode address to `nullptr` and assuming that they were located at the
end of the instruction stream (which was true for dx-generated code, but not
required by the spec). I got around this by inserting MFLOW_FALLTHROUGH entries
to represent these opcodes. This simplified other code: for example, now that
every valid address offset had a corresponding MIE, `insert_mentry_before` no
longer needed to check if the insertion point was a valid MIE or a nullptr.

Note that this diff results in a slight behavioral change: previously, we would
not insert MFLOW_POSITION/MFLOW_DEBUG items that were adjacent to
nops/pseudo-opcodes. Now we do. Because of the way we construct the CFG, the
extra entries make Dedupblocks not dedup some blocks. For example:

Before:

```
B0: MFLOW_OPCODE: goto
B1: MFLOW_OPCODE: goto
```

After:

```
B0: MFLOW_OPCODE: goto
B1: MFLOW_POSITION
B2: MFLOW_OPCODE: goto
```

Now it appears that B1 "falls through" to B2, even though they should really
be considered as a single basic block.

In any case, I don't think this has a significant size effect (will let stats
bot verify), so I'm not fixing it.

Reviewed By: justinjhendrick

Differential Revision: D5600561

fbshipit-source-id: 52fa12b340143a8982f48612f4dc4ce17d2a4f92