Add StringSimplification Pass
Summary:
This commit adds the `StringSimplification` pass which is an extended form of constant propagation for strings.  It attempts to remove unnecessary string builders by computing their values beforehand.

This works locally, so it should be run after global constant propagation and a pass of `LocalDce` should be run after to clean up any unnecessary constant strings (such as the ones before the appending).

Reviewed By: minjang

Differential Revision: D5455986

fbshipit-source-id: c6173a1bdbf7b9cd080bbf218ab0e45ccf8e423d