Adds whitelist for classes with no size limit per method
Summary:
As title.

To be used in next diffs. If we want to dedup more (for type erasure), we are
probably going to perform inlining + some dedup. This requires the limit to be
removed, as we've already seen that it hits this limit.

In case we see other errors because of removing this limit, will rethink.

For now, it should only be used for specific cases.

Reviewed By: thezhangwei

Differential Revision: D5628802

fbshipit-source-id: 79830bf60301d3783882d4846f70034e6784b79c