Remove inline_tail_call's hacks around lack of regalloc
Summary:
This diff removes the code that tries to keep the parameter registers of the caller at the end of the frame, as well as the checks that prevent bridgepass from inlining methods if the final result will have more than 16 registers.

Initially, I thought we could remove `inline_tail_call` and replace its uses with `inline_method`. However, while inline_tail_call simply remaps the parameter registers in the callee, inline_method has to insert moves to map the argument registers in the caller to the param registers in the callee (since it doesn't know if those argument registers will be live-out after the invoke). These extra move instructions confuse code that is looking for exact sequences of opcodes. In particular, inserting extra move instructions in BridgePass' inlining will cause SynthPass to match fewer methods. As such, I am keeping `inline_tail_call` around for now, at least until we fix the exact opcode matches to be more flexible.

Reviewed By: dariorussi

Differential Revision: D5567130

fbshipit-source-id: 10e1dc95af81c0edbc797c99031d88a17bdf7996