Register allocation: Fix bug in live range splitting
Summary:
This bug didn't appear now, but it appeared after change in D5579111.
A multi-branch instruction might be the last use of the register we want to split around, then normally the load move will be inserted after the instruction, then the load move will end up in the block that Fallthrough, but it should actually get inserted for each block that this branch goes to. Before I dealt with this case for conditional branch, but actually sparse_switch and packed_switch can cause problem too.

Reviewed By: int3

Differential Revision: D5633876

fbshipit-source-id: 185051d3d191c27b5a18753e9e820c9bb52e716e