Do not add edge for unreachable blocks in cfg.
Summary:
This commit is fixing the problem raised in D5623546. There are some unreachable blocks in cfg when doing register allocation. Before RegAlloc called allocate(), it called remove_unreachable_blocks() to remove unreachable blocks. Although the blocks that contains position MethodItemEntry and try_start try_end MethodItemEntry are still there, the edges to their successor blocks are removed, so this should not cause problem for finding immediate dominator. However if we do live_range_splitting before finding immediate dominator, the cfg was rebuilt because there were some new blocks added during splitting and we wanted to have a correct cfg. So the edges to those unreachable blocks are reconnected, then finding immediate dominator for the cfg will be problematic:

   B0     B2(unreachable)
     \       /
       B1: finding immediate dominator for B1 will result in infinite loop in D5623546.

This commit removes the edges between unreachable blocks and their successor blocks in build_cfg(), so that the unreachable blocks are isolated from reachable blocks in cfg.

Reviewed By: int3

Differential Revision: D5628536

fbshipit-source-id: 0918e5cc9a3ca41cca951a096e305128555febcb