Code to check whether an oatfile was created by oatmeal.
Summary:
Code to check whether an oatfile was created by oatmeal,
and corresponding trace tag "oatmeal". If ART rejects an oat file
for any reason, it silently deletes it and creates a new oatfile.
Therefore, we can't simply assume that the loaded oat file was
the one we created, and need to track this failure after the fact.

Reviewed By: kgsharma, maxmg22

Differential Revision: D5605616

fbshipit-source-id: 82c68cfa8c5a65d1669a7df40fcd25de0a8c68e8