Support for S-expressions
Summary:
S-expressions are a compact, efficient and readable format for serializing complex data structures (abstract semantics, program invariants, internal representations, etc.). This library provides support for S-expressions, including a parser and pattern-matching functionalities.

This is a header-only library, because it's good practice (Boost is header-only for the most part) and it also allows us to define templates.

Reviewed By: int3

Differential Revision: D5710042

fbshipit-source-id: 93a10774a6a1053c032400fcbbd1d008f2a8087d