Fix same builder initialized multiple times.
Summary:
We were not treating correctly the case where we have multiple builders (of the
same type) initialized. In order to properly treat those, we need a better
analysis of the builder instances (which we are not doing now). Leaving this as
a TODO given there aren't many cases, so we are skipping those for now.

Reviewed By: thezhangwei, justinjhendrick

Differential Revision: D5726883

fbshipit-source-id: dc4e049049fff8c93dedddefe6f30bd96ee1edcb