A type checker for the IR
Summary:
This is a fast type checker for the IR. It infers the type of all registers at each instruction and detects common errors, like fetching a reference from a register that holds an integer or incorrectly accessing the components of a wide register.

It is mostly meant to be used as a sanity check at the end of an optimization pass. In particular, the type checker doesn't try to replicate all the features of the Android verifier.

The inferred types are also available for use by any analysis/optimization pass that requires type information.

Reviewed By: int3, justinjhendrick

Differential Revision: D5773279

fbshipit-source-id: 7760d0c1c79beda8f9de16159c720087302eca31