Fixed the normalization operation in AbstractDomainScaffolding
Summary:
The normalization operation of AbstractDomainScaffolding is only meant to keep the m_kind flag and the internal representation in sync after an operation that produces Top. Bottom should always be treated as a special case in all operations of the abstract domain. I corrected the implementation of normalize() and I fixed some operations in HashedAbstractEnvironment and PatriciaTreeMapAbstractEnvironment that relied on the normalization to handle the case where the result is Bottom.

This bug doesn't affect the correctness of the existing analyses. Its main effect is to turn Bottom invariants into any non-Bottom value, even Top. This creates noise that can contaminate other invariants and negatively impact the precision of an analysis.

Reviewed By: minjang

Differential Revision: D5822159

fbshipit-source-id: 5f34ce9a03daac2eb5a8ce83ea9f86b3f3b9750f