Fix equivalent method + re-enable queries
Summary:
As title.

Before, for checking methods are equal, we were only looking at the `MFLOW_OPCODE` entries. This should be enough in most of the cases, except when there is a switch, given we should also check that the cases are the same. For this, added `structural_equals` to the IRCode.

Reviewed By: thezhangwei

Differential Revision: D5852761

fbshipit-source-id: 982d3a79907437f4d8e16a993f612c46bc9724be