Enable the IR type checker in the configuration file
Summary: The IR type checker can now be controlled from the JSON configuration file. When enabled, it runs after all optimization passes. In the presence of a type error, it is possible to run the type checker after each pass in order to identify the pass that causes the inconsistency.

Reviewed By: int3

Differential Revision: D5870730

fbshipit-source-id: b0c1b58ad98557a63e336f589f58ef068292fd38