Added support for polymorphic constants to the IR type checker
Summary:
The current version of the IR type checker considers constants to be polymorphic, which isn't the behavior of the Android verifier.

Now, the default behavior of the type checker complies with the Android verifier. Polymorphic constants can be enabled using a flag. This flag is automatically set when verify_none_mode is enabled.

Reviewed By: justinjhendrick

Differential Revision: D5909417

fbshipit-source-id: 8201960c11da9003d500d30503ced89181171806