Check deobfuscated name when matching the whitelisted / blacklisted interfaces
Summary:
As title.

When checking if an interface is whitelisted or blacklisted, check the
deobfuscated name. If Proguard class name renaming is not enabled, the
deobfuscated name is the same as the name.

For most of the apps, Proguard class name renaming is disabled, while WhatsApp
uses it. With this change, we can blacklist specific interfaces for WhatsApp too.

Reviewed By: justinjhendrick

Differential Revision: D5929106

fbshipit-source-id: 64186619adfefa56bdde886a2232cda10708b235