option to not mark manifest/xml layout as reachable
Summary: Proguard rules should reasonably protect against the kind of changes this was attempting to cover; the Android Proguard config that ships with the SDK certainly does. This redex code seems to exist for legacy reasons; e.g. from the time before Redex was able to parse Proguard configs and we had to do something overly conservative. Making it an option since some apps may have grown to depend on this overly conservative view.

Reviewed By: dariorussi

Differential Revision: D5897419

fbshipit-source-id: e5b4dea20457c87e46ee63814e8de1f6531a9330