Remove unreferenced interfaces
Summary:
If an interface is internal and never referenced in any form through the code base we can safely remove the interface.
We got almost 2K such interfaces so it seems good to do.
Of course it's possible and likely that other passes were already doing work to remove methods in the interfaces and more but we knew from TypeErasure and PG deprecation that there are a bunch of unremoved ones

build[automation_fbandroid_redexunstable_]

Reviewed By: thezhangwei, emma0303

Differential Revision: D5895305

fbshipit-source-id: b6d497bc2a0c51ba121291a2ff7129bfbead0abe