Delay file open in oatmeal to accommodate for V124
Summary:
Up until Oreo, all the OAT metadata and DEX bytecode would be place in a single file, the famous `everything.oat`.
With Oreo, we now have `.odex` for OAT metadata and `.vdex` for bytecode so we have to potentially open two files to
be able to parse the bytecode content.  This set of changes setup oatmeal to be able to open `odex` and `vdex` for V124.

Support for V124 parsing comes in another diff.

Reviewed By: mlogan

Differential Revision: D5983299

fbshipit-source-id: 3a79a6563d6324729aab5e42fdf716a7c6d481cd