typechecker: Check for structural constraints as well
Summary:
I've encountered bugs with the following constraints being violated, and I
thought it would be useful to check for them as well:

* load-param opcodes should all be at the start of the method
* move-result opcodes should always be prefixed by an appropriate instruction

When I add move-result-pseudo instructions, I will add checks for them here as
well.

Additionally, I added a print of the full method bytecode upon encountering an
error. IME it makes it much quicker to debug the errors.

Remark: Maybe the typechecker should be renamed to be a verifier, now that this
diff makes it do more?

Reviewed By: arnaudvenet

Differential Revision: D5992519

fbshipit-source-id: 89d95d5d370d01c4764945d7708fda75d66ce119