Refactor and convert instruction selection away from a pass
Summary:
I initially thought it would be good as a pass so that we could reuse the
existing stats reporting infrastructure. However, it led to a number of awkward
consequences:

* Every single redex config file needed to include InstructionSelectionPass
* The typechecker needed to handle 2addr opcodes because instruction selection
  would reintroduce them

Going ahead, I would like the instruction selection step to do the conversion
of IRInstructions to DexInstructions. This would make it even more difficult for
the typechecker to run afterward. As such, I am giving it the special treatment
it deserves and distinguishing it from an ordinary pass.

Minor refactoring details:

I've also renamed it from 'instruction selection' to 'instruction lowering'
because I think eventually the majority of its work will be in doing a 1-1
conversion between IR and Dex instructions. The amount of selecting between
different choices will be limited. I've also eliminated the use of a class
in favor of top-level functions since we do not seem to need any state.
Finally, I moved some functions under an `impl` namespace to make it
clear that other passes should not attempt to use them.

Reviewed By: justinjhendrick, maxmg22

Differential Revision: D5995438

fbshipit-source-id: fd277ab9f43cfc4a5a4cbb45ba8f08065033eda9