Have instruction lowering convert IR to DexInstructions
Summary:
Prior to this diff, instruction lowering would map IR to IR. This meant that it was
limited in the kinds of transformations it could represent -- the input and output
had to both share the same semantics. This diff allows us to do more in the lowering
step. (The first thing will be to implement move-result-pseudo).

The current approach is hacky and non-ideal. I am adding another tag -- MFLOW_DEX_OPCODE --
to our MethodItemEntry union in order to store DexInstructions there. Because IRInstruction
and DexInstruction have very similar APIs, this is actually quite confusing and bug-prone.
On the bright side, MFLOW_DEX_OPCODE should never be used in any pass, so most redex
development will not need to grapple with this.

Going ahead, I think a good approach is to have lowering generate a new data structure altogether
(i.e. no reusing of FatMethod). Perhaps that can tie in with justinjhendrick's work on CFG linearization.

Reviewed By: justinjhendrick

Differential Revision: D5995456

fbshipit-source-id: 8da28d585ab2d677b10896577b4edd5a5e5b0925